requirements.txt
cupy-cuda12x
triton
scipy
numpy
pytest
torch